I added monsters, to maps, added exit areas

+++ the_new_rpg/src/Rock.py
@@ -20,6 +20,6 @@ class Rock():
         self.x *= s
         self.y *= s
         #scale base pic to right size
-        self.base = pygame.transform.scale(self.base, [self.base.get_width()*s, self.base.get_height*s])
+        self.base = pygame.transform.scale(self.base, [int(self.base.get_width()*s), int(self.base.get_height()*s)])
         self.w = self.base.get_width()
         self.h = self.base.get_height()
